Removed all imports.
This package is a big collection of miscellaneous stuff, so it's common to want to use only a small piece. If all the dependencies are imported, this requires users of the package to install all the dependencies, even the ones that will never be used. Now it's the user's job to install the dependencies that he'll actually need.